Привязка программы к направлению

+++ database/migrations/45_create_linking_the_program_to_the_direction_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('linking_the_program_to_the_direction', function (Blueprint $table) {
-            $table->integer('Program_ID');
-            $table->integer('direction_ID');
+            $table->unsignedBigInteger('Program_ID');
+            $table->unsignedBigInteger('direction_ID');
             $table->primary(['Program_ID', 'direction_ID']);
         });
     }
